Add benchmark_test.go, example_test.go
